+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   trigger function helper.
 * Remove `Read` instance of `IndexMethod`.
 * Allow using nonreserved keywords, e.g. `timestamp`, in table indices.
+* Add support for the `NUMERIC` column type.
 
 # hpqtypes-extras-1.19.0.0 (2025-11-27)
 * Compatibility with `hpqtypes` >= 0.13.0.0.